start clean up of index fund;

+++ contracts/core/struct.sol
@@ -403,7 +403,7 @@ library AngelCoreStruct {
         uint256 id;
         string name;
         string description;
-        uint256[] members;
+        uint32[] members;
         bool rotatingFund; // set a fund as a rotating fund
         //Fund Specific: over-riding SC level setting to handle a fixed split value
         // Defines the % to split off into liquid account, and if defined overrides all other splits
@@ -418,7 +418,7 @@ library AngelCoreStruct {
     }
 
     struct BeneficiaryData {
-        uint256 id;
+        uint32 id;
         address addr;
     }
